Allow injecting validate.js options
It's now possible to set up a static `validatorOptions` object in the stimulus
controller. This object will be used to pass any extra options accepted by validate.js.

Eg, validate.js has an option to skip prepending the error msgs with the attr name: `fullMessages`.
When setting this option to false, instead of an error msg being "Foo can't be blank", it'll be
"can't be blank".

It's now easy to pass this option by setting:

   static rules = {...}
   static validatorOptions = { fullMessages: false }

---